admin panel category add/delete/edit

+++ products/models.py
@@ -37,7 +37,7 @@ class Product(models.Model):
     product_brand  = models.CharField(max_length=100, default="No Brand")
     product_category = models.ForeignKey(Subcategory, on_delete=models.CASCADE, related_name='products')
     product_slug = models.SlugField(unique=True, null=True, blank=True)
-    product_price = models.IntegerField()
+    product_price = models.DecimalField(max_digits=15, decimal_places=2)
     product_description = models.TextField()
     product_quantity = models.IntegerField(null=True)
     product_sold_quantity = models.IntegerField(default=0)
